Use read and write thread in AIO TCP client

+++ universe/common/src/main/java/com/wfwgyy/imsa/common/AppConsts.java
@@ -24,11 +24,11 @@ public class AppConsts {
 	public final static int MT_MSG_V5 = 5;
 	// Redis服务器
 	public final static String REDIS_HOST = SERVER_IP;
-	public final static short REDIS_PORT = 6379;
+	public final static int REDIS_PORT = 6379;
 	// 消息服务器相关配置
 	public final static String PLATO_HOST = SERVER_IP;
-	public final static short PLATO_PORT = 8089;
+	public final static int PLATO_PORT = 8089;
 	// 门户Facade相关
 	public final static String FACADE_HOST = SERVER_IP;
-	public final static short FACADE_PORT = 8088;
+	public final static int FACADE_PORT = 8088;
 }
